moved the DAO Object file to a new folder called text_file that is nested inside data_access
